Fixed bug in verify on 32-bit systems
A bug in verify occurring on 32-bit systems, due to the use of
size_t to represent file offsets, was fixed.

+++ trunk/src/verify.c
@@ -663,7 +663,7 @@ calc_chksums(int fd, unsigned char *initsum, unsigned char *sum,
     const struct aiocb *cbp;
     EVP_MD_CTX ctx, initctx;
     int err;
-    size_t flen = 0, initrem = 512;
+    off_t flen = 0, initrem = 512;
     struct aiocb cb;
 
     static char buf1[BUFSIZE], buf2[BUFSIZE];
